+++ src/com/ansorgit/plugins/bash/lang/lexer/HeredocLexingState.java
@@ -17,37 +17,87 @@ package com.ansorgit.plugins.bash.lang.lexer;
 
 import com.ansorgit.plugins.bash.lang.util.HeredocSharedImpl;
 import com.google.common.collect.Lists;
-import com.intellij.openapi.util.Pair;
 
 import java.util.LinkedList;
 
 /**
  * Heredoc lexing state used in the lexer
  */
-class HeredocLexingState {
-    private final LinkedList<Pair<String, Boolean>> expectedHeredocs = Lists.newLinkedList();
+final class HeredocLexingState {
+    private final LinkedList<HeredocMarkerInfo> expectedHeredocs = Lists.newLinkedList();
 
-    boolean isNextHeredocMarker(String marker) {
-        return !expectedHeredocs.isEmpty() && expectedHeredocs.peekFirst().first.equals(HeredocSharedImpl.cleanMarker(marker));
+    public boolean isEmpty() {
+        return expectedHeredocs.isEmpty();
     }
 
-    void pushHeredocMarker(String marker) {
-        expectedHeredocs.add(Pair.create(HeredocSharedImpl.cleanMarker(marker), HeredocSharedImpl.isEvaluatingMarker(marker)));
+    boolean isNextMarker(CharSequence markerText) {
+        return !expectedHeredocs.isEmpty() && expectedHeredocs.peekFirst().nameEquals(markerText);
     }
 
     boolean isExpectingEvaluatingHeredoc() {
-        return !expectedHeredocs.isEmpty() && expectedHeredocs.peekFirst().second;
+        if (isEmpty()) {
+            throw new IllegalStateException("isIgnoringTabs called on an empty marker stack");
+        }
+
+        return !expectedHeredocs.isEmpty() && expectedHeredocs.peekFirst().evaluating;
     }
 
-    void popHeredocMarker(String marker) {
-        if (!isNextHeredocMarker(HeredocSharedImpl.cleanMarker(marker))) {
+    boolean isIgnoringTabs() {
+        if (isEmpty()) {
+            throw new IllegalStateException("isIgnoringTabs called on an empty marker stack");
+        }
+
+        return !expectedHeredocs.isEmpty() && expectedHeredocs.peekFirst().ignoreLeadingTabs;
+    }
+
+    void pushMarker(CharSequence marker, boolean ignoreTabs) {
+        expectedHeredocs.add(new HeredocMarkerInfo(marker, ignoreTabs));
+    }
+
+    void popMarker(CharSequence marker) {
+        if (!isNextMarker(HeredocSharedImpl.cleanMarker(marker.toString(), false))) {
             throw new IllegalStateException("Heredoc marker isn't expected to be removed: " + marker);
         }
 
         expectedHeredocs.removeFirst();
     }
 
-    public boolean isEmpty() {
-        return expectedHeredocs.isEmpty();
+    private static class HeredocMarkerInfo {
+        final boolean ignoreLeadingTabs;
+        final boolean evaluating;
+        final CharSequence markerName;
+
+        HeredocMarkerInfo(CharSequence markerText, boolean ignoreLeadingTabs) {
+            String markerTextString = markerText.toString();
+
+            this.markerName = HeredocSharedImpl.cleanMarker(markerTextString, ignoreLeadingTabs);
+            this.evaluating = HeredocSharedImpl.isEvaluatingMarker(markerTextString);
+            this.ignoreLeadingTabs = ignoreLeadingTabs;
+        }
+
+        boolean nameEquals(CharSequence markerText) {
+            return this.markerName.equals(HeredocSharedImpl.cleanMarker(markerText.toString(), ignoreLeadingTabs));
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+
+            HeredocMarkerInfo that = (HeredocMarkerInfo) o;
+
+            if (ignoreLeadingTabs != that.ignoreLeadingTabs) return false;
+            if (evaluating != that.evaluating) return false;
+            return markerName != null ? markerName.equals(that.markerName) : that.markerName == null;
+
+        }
+
+        @Override
+        public int hashCode() {
+            int result = (ignoreLeadingTabs ? 1 : 0);
+            result = 31 * result + (evaluating ? 1 : 0);
+            result = 31 * result + (markerName != null ? markerName.hashCode() : 0);
+            return result;
+        }
     }
 }