+++ src/com/ansorgit/plugins/bash/lang/psi/util/BashPsiElementFactory.java
@@ -31,6 +31,7 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
 import com.intellij.psi.PsiFileFactory;
 import com.intellij.psi.util.PsiTreeUtil;
+import org.apache.commons.lang.StringUtils;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -123,8 +124,8 @@ public class BashPsiElementFactory {
         return PsiTreeUtil.findChildOfType(createDummyBashFile(project, data), BashHereDocStartMarker.class);
     }
 
-    public static PsiElement createHeredocEndMarker(Project project, String name) {
-        String data = String.format("cat << %s\n%s", name, name);
+    public static PsiElement createHeredocEndMarker(Project project, String name, int leadingTabs) {
+        String data = String.format("cat <<- %s\n%s", name, StringUtils.repeat("\t", leadingTabs) + name);
         return PsiTreeUtil.findChildOfType(createDummyBashFile(project, data), BashHereDocEndMarker.class);
     }