+++ src/com/ansorgit/plugins/bash/lang/util/HeredocSharedImpl.java
@@ -21,29 +21,29 @@ import org.jetbrains.annotations.NotNull;
 
 /**
  * Shared code for the Heredoc handling.
+ *
+ * @author jansorg
  */
-public final class HeredocSharedImpl {
-    private HeredocSharedImpl() {
-    }
-
-    public static String cleanMarker(String marker) {
+public class HeredocSharedImpl {
+    public static String cleanMarker(String marker, boolean ignoredLeadingTabs) {
         String markerText = trimNewline(marker);
         if (markerText.equals("$")) {
             return markerText;
         }
 
-        int start = startMarkerTextOffset(markerText);
-        int end = endMarkerTextOffset(markerText);
+        Pair<Integer, Integer> offsets = getStartEndOffsets(markerText, ignoredLeadingTabs);
+        int start = offsets.first;
+        int end = offsets.second;
 
         return end <= markerText.length() && start < end ? markerText.substring(start, end) : marker;
     }
 
-    public static int startMarkerTextOffset(String markerText) {
-        return getStartEndOffsets(markerText).first;
+    public static int startMarkerTextOffset(String markerText, boolean ignoredLeadingTabs) {
+        return getStartEndOffsets(markerText, ignoredLeadingTabs).first;
     }
 
     public static int endMarkerTextOffset(String markerText) {
-        return getStartEndOffsets(markerText).second;
+        return getStartEndOffsets(markerText, false).second;
     }
 
     public static boolean isEvaluatingMarker(String marker) {
@@ -52,20 +52,21 @@ public final class HeredocSharedImpl {
         return !markerText.startsWith("\"") && !markerText.startsWith("'") && !markerText.startsWith("\\") && !markerText.startsWith("$");
     }
 
+    private static String trimNewline(String marker) {
+        return StringUtils.removeEnd(marker, "\n");
+    }
+
     public static String wrapMarker(String newName, String originalMarker) {
-        int start = startMarkerTextOffset(originalMarker);
-        int end = endMarkerTextOffset(originalMarker);
+        Pair<Integer, Integer> offsets = getStartEndOffsets(originalMarker, true);
+        int start = offsets.first;
+        int end = offsets.second;
 
         return (end <= originalMarker.length() && start < end)
                 ? originalMarker.substring(0, start) + newName + originalMarker.substring(end)
                 : newName;
     }
 
-    private static String trimNewline(String marker) {
-        return StringUtils.removeEnd(marker, "\n");
-    }
-
-    private static Pair<Integer, Integer> getStartEndOffsets(@NotNull String markerText) {
+    private static Pair<Integer, Integer> getStartEndOffsets(@NotNull String markerText, boolean ignoredLeadingTabs) {
         if (markerText.isEmpty()) {
             return Pair.create(0, 0);
         }
@@ -82,7 +83,11 @@ public final class HeredocSharedImpl {
         int start = 0;
         int end = length - 1;
 
-        if (markerText.charAt(start) == '$' && length > 2 && (markerText.charAt(start + 1) == '"' || markerText.charAt(end) == '\'')) {
+        while (ignoredLeadingTabs && start < length && markerText.charAt(start) == '\t') {
+            start++;
+        }
+
+        if (markerText.charAt(start) == '$' && length > (start + 2) && (markerText.charAt(start + 1) == '"' || markerText.charAt(end) == '\'')) {
             start++;
             length--;
         }
@@ -94,6 +99,7 @@ public final class HeredocSharedImpl {
         if (length > 0 && (markerText.charAt(start) == '\'' || markerText.charAt(start) == '"') && markerText.charAt(end) == markerText.charAt(start)) {
             start++;
             end--;
+            length -= 2;
         }
 
         return Pair.create(start, end + 1);